classes transform: use _createClass instead of prototype
